third_party/WebKit/LayoutTests/http/tests/devtools/bindings/contentscripts-navigator-multiple-frames.html

<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/debugger-test.js"></script>
<script src="../../inspector/bindings/bindings-test.js"></script>

<script>

async function test() {
  var contentScriptsNavigator = new Sources.ContentScriptsNavigatorView();
  contentScriptsNavigator.show(UI.inspectorView.element);

  TestRunner.markStep('initialWorkspace');
  SourcesTestRunner.dumpNavigatorView(contentScriptsNavigator, false);

  TestRunner.markStep('attachFrame1');
  await BindingsTestRunner.attachFrame('frame1', './resources/contentscript-frame.html', '_test_attachFrame1.js'),
      SourcesTestRunner.dumpNavigatorView(contentScriptsNavigator, false);

  TestRunner.markStep('attachFrame2');
  await BindingsTestRunner.attachFrame('frame2', './resources/contentscript-frame.html', '_test_attachFrame2.js'),
      SourcesTestRunner.dumpNavigatorView(contentScriptsNavigator, false);

  TestRunner.markStep('detachFrame1');
  await BindingsTestRunner.detachFrame('frame1', '_test_detachFrame1.js');
  SourcesTestRunner.dumpNavigatorView(contentScriptsNavigator, false);

  TestRunner.markStep('detachFrame2');
  await BindingsTestRunner.detachFrame('frame2', '_test_detachFrame2.js');
  SourcesTestRunner.dumpNavigatorView(contentScriptsNavigator, false);

  TestRunner.completeTest();
}

</script>
</head>
<body onload="runTest()">
<p>
Verify that SourceMap bindings are generating UISourceCodes properly.
</p>
</body>
