third_party/WebKit/LayoutTests/http/tests/devtools/bindings/contentscripts-navigator-multiple-frames-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verify that SourceMap bindings are generating UISourceCodes properly.


Running: initialWorkspace

Running: attachFrame1
top
  (no domain)
    content-script-42.js

Running: attachFrame2
top
  (no domain)
    content-script-42.js

Running: detachFrame1
top
  (no domain)
    content-script-42.js

Running: detachFrame2
top
  (no domain)
    content-script-42.js
